git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@5436 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickWand/magick-image.h
@@ -77,7 +77,7 @@ extern WandExport MagickBooleanType
   MagickAdaptiveSharpenImage(MagickWand *,const double,const double,
     const double),
   MagickAdaptiveThresholdImage(MagickWand *,const size_t,const size_t,
-    const ssize_t),
+    const double),
   MagickAddImage(MagickWand *,const MagickWand *),
   MagickAddNoiseImage(MagickWand *,const NoiseType),
   MagickAffineTransformImage(MagickWand *,const DrawingWand *),
